# perf(storage): add foyer memory cache to block cache bench